requirements.txt
inference
supervision
python-dotenv
playsound